+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -14,6 +14,13 @@ and moved to `BuildModelOptions`, `CreateComposedModelOptions` and `CopyAuthoriz
 
 ### Other Changes
 
+## 3.1.10 (2022-04-08)
+
+### Other Changes
+#### Dependency Updates
+- Upgraded `azure-core` from `1.26.0` to version `1.27.0`.
+- Upgraded `azure-core-http-netty` from `1.11.8` to version `1.11.9`.
+
 ## 3.1.9 (2022-03-08)
 ### Bug Fixes
 - Fixed referencing selection mark form element from FormField element.